fix: team attendance scan
